COMPAT: fixup decimal extension for indexing compat (#19882)

+++ pandas/tests/extension/base/getitem.py
@@ -18,7 +18,8 @@ class BaseGetitemTests(BaseExtensionTests):
         self.assert_series_equal(result, expected)
 
     def test_iloc_frame(self, data):
-        df = pd.DataFrame({"A": data, 'B': np.arange(len(data))})
+        df = pd.DataFrame({"A": data, 'B':
+                           np.arange(len(data), dtype='int64')})
         expected = pd.DataFrame({"A": data[:4]})
 
         # slice -> frame
@@ -49,7 +50,8 @@ class BaseGetitemTests(BaseExtensionTests):
         self.assert_series_equal(result, expected)
 
     def test_loc_frame(self, data):
-        df = pd.DataFrame({"A": data, 'B': np.arange(len(data))})
+        df = pd.DataFrame({"A": data,
+                           'B': np.arange(len(data), dtype='int64')})
         expected = pd.DataFrame({"A": data[:4]})
 
         # slice -> frame
